+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added color and intensity customization for Decals.
 - Added a history rejection criterion based on if the pixel was moving in world space (case 1302392).
 - Added the default quality settings to the HDRP asset for RTAO, RTR and RTGI (case 1304370).
-- Added a Force Forward Emissive optoin for Lit Material to force to render the Emissive contribution in a separate forward pass when the Lit Material is in Deferred Lit shader Mode.
+- Added a Force Forward Emissive option for Lit Material that forces the Emissive contribution to render in a separate forward pass when the Lit Material is in Deferred Lit shader Mode.
 
 ### Fixed
 - Fixed an exception when opening the color picker in the material UI (case 1307143).